IoT/requirements.txt
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
Pillow==9.5.0
Pillow==10.1.0
ultralytics==8.0.136
